requirements_enhanced.txt
# Enhanced Jarvis 2.0 Dependencies - Continuous Conversation & AI Integration

# Core existing dependencies
audioop-lts==0.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comtypes==1.4.11
decorator==5.2.1
docopt==0.6.2
Flask==3.1.1
future==1.0.0
geocoder==1.38.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Js2Py==0.74
MarkupSafe==3.0.2
MouseInfo==0.1.3
packaging==25.0
pillow==11.3.0
pipwin==0.5.2
psutil==7.0.0
PyAudio
PyAutoGUI==0.9.54
pycaw==20240210
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyperclip==1.9.0
pypiwin32==223
PyPrind==2.11.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
pyttsx3==2.98
pytweening==1.2.0
pywhatkit==5.4
pywin32
ratelim==0.1.6
requests==2.32.4
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
SpeechRecognition==3.14.3
speedtest-cli==2.1.3
standard-aifc==3.13.0
standard-chunk==3.13.0
tk==0.1.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.4
wheel==0.45.1
wikipedia==1.4.0

# Enhanced Speech Processing
openai-whisper==20231117
transformers==4.53.0
torch==2.8.0
torchaudio==2.1.0

# Advanced NLP and Machine Learning
scikit-learn==1.5.0
nltk==3.9
sentence-transformers==2.2.2
google-generativeai==0.8.3

# Database and Storage
sqlalchemy==2.0.23

# Enhanced Audio Processing
soundfile==0.12.1
librosa==0.10.1
webrtcvad==2.0.10

# Voice Activity Detection and Audio Processing
pyaudio-fork==0.2.13
speechrecognition-whisper==1.0.0

# Additional utilities for conversation management
python-dateutil==2.8.2
asyncio-throttle==1.0.2
python-dotenv